+++ tests/test_function/test_parse_resource.py
@@ -34,6 +34,20 @@ def test_parse_list_keys() -> None:
     assert_that(result).is_equal_to(expected_result)
 
 
+def test_parse_pick_zones() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "pick_zones"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BicepParser(file_path).json()
+
+    # then
+    print(json.dumps(result, indent=2))
+    assert_that(result).is_equal_to(expected_result)
+
+
 def test_parse_reference() -> None:
     # given
     sub_dir_path = EXAMPLES_DIR / "reference"